Add WSEventType union type, begin alphabetizing types
Given that WSEventType is used in an array, a union type is the best choice here because the array used will offer autocompletion for each individual event when passing these events to ClientOptions. I intend to use a union type for permission strings as well, internally, to offer autocompletion for PermissionResolvables as well.