Add TextBasedChannel & PartialTextBasedChannel mixin model/support
I've been waiting to be able to model this for so long. TypeScript 2.2 is my savior. We've never been able to  get around having to manually keep definitions for all these methods on each class that uses them, thus having to update them in like 6 different places. Thing of the past, that.